# Improve performance of Tensorflow and PyTorch modules