--- a/3rdparty/CMake/FindPytorch.cmake
+++ b/3rdparty/CMake/FindPytorch.cmake
@@ -20,21 +20,27 @@ if(NOT Pytorch_FOUND)
 
     message(STATUS "Getting PyTorch properties ...")
 
-    # Get Pytorch_VERSION
-    execute_process(
-        COMMAND ${PYTHON_EXECUTABLE} "-c"
-                "import torch; print(torch.__version__, end='')"
-        OUTPUT_VARIABLE Pytorch_VERSION)
-
-    # Get Pytorch_ROOT
+    set(Pytorch_FETCH_PROPERTIES
+        "import os"
+        "import torch"
+        "print(torch.__version__, end=';')"
+        "print(os.path.dirname(torch.__file__), end=';')"
+        "print(torch._C._GLIBCXX_USE_CXX11_ABI, end=';')"
+    )
     execute_process(
         COMMAND
-            ${PYTHON_EXECUTABLE} "-c"
-            "import os; import torch; print(os.path.dirname(torch.__file__), end='')"
-        OUTPUT_VARIABLE Pytorch_ROOT)
+            ${PYTHON_EXECUTABLE} "-c" "${Pytorch_FETCH_PROPERTIES}"
+        OUTPUT_VARIABLE PyTorch_PROPERTIES
+    )
+
+    list(GET PyTorch_PROPERTIES 0 Pytorch_VERSION)
+    list(GET PyTorch_PROPERTIES 1 Pytorch_ROOT)
+    list(GET PyTorch_PROPERTIES 2 Pytorch_CXX11_ABI)
+
+    unset(PyTorch_PROPERTIES)
 
     # Use the cmake config provided by torch
-    find_package(Torch REQUIRED PATHS "${Pytorch_ROOT}/share/cmake/Torch"
+    find_package(Torch REQUIRED PATHS "${Pytorch_ROOT}"
                  NO_DEFAULT_PATH)
 
     if(BUILD_CUDA_MODULE)
@@ -62,14 +68,6 @@ if(NOT Pytorch_FOUND)
         set_target_properties( torch_cuda PROPERTIES INTERFACE_COMPILE_OPTIONS "" )
         set_target_properties( torch_cpu PROPERTIES INTERFACE_COMPILE_OPTIONS "" )
     endif()
-
-    # Get Pytorch_CXX11_ABI: True/False
-    execute_process(
-        COMMAND
-            ${PYTHON_EXECUTABLE} "-c"
-            "import torch; print(torch._C._GLIBCXX_USE_CXX11_ABI, end='')"
-        OUTPUT_VARIABLE Pytorch_CXX11_ABI
-    )
 endif()
 
 message(STATUS "PyTorch         version: ${Pytorch_VERSION}")